Rename launch_pad to launch_app
