+++ README.md
@@ -132,7 +132,7 @@ If the SQLite file becomes corrupted, you can rebuild it directly from the
 Launch Pad utility:
 
 ```bash
-python launch_pad.py
+python launch_app.py
 ```
 
 Open the **Operations** menu and choose **Recover Database**. This deletes the
