+++ tests/test_launch_app_recover.py
@@ -21,9 +21,9 @@ sys.modules.setdefault("rich.text", dummy_text)
 import pytest
 
 try:
-    import launch_pad
+    import launch_app
 except Exception:
-    pytest.skip("launch_pad module unavailable", allow_module_level=True)
+    pytest.skip("launch_app module unavailable", allow_module_level=True)
 
 
 def test_operations_menu_recover(monkeypatch):
@@ -54,9 +54,9 @@ def test_operations_menu_recover(monkeypatch):
 
     inputs = iter(["3", "", "b"])
     monkeypatch.setattr(builtins, "input", lambda _: next(inputs))
-    monkeypatch.setattr(launch_pad, "clear_screen", lambda: None)
-    monkeypatch.setattr(launch_pad, "DataLocker", DummyLocker)
+    monkeypatch.setattr(launch_app, "clear_screen", lambda: None)
+    monkeypatch.setattr(launch_app, "DataLocker", DummyLocker)
 
-    launch_pad.operations_menu()
+    launch_app.operations_menu()
 
     assert called["recover"] is True
